
\name{Apoderoides-internal}
\alias{deleteSubFunc}
\alias{getRankMRCA}
\alias{getAllRankNames}
\alias{findUpperNode}
\alias{is.monophyleticByRank}
\alias{getAllCentroids_C}
\alias{calcIntScore_C}
\alias{calcOutScore_C}
\alias{extractOTUbyRankName_C}
\alias{findSubTips_C}
\alias{getTipNo_C}
\alias{getRankCentroid_C}
\title{Internal Apoderoides Functions}
\description{
 Internal Apoderoides functions
}
\details{
These are not to be called by the user.
}
\value{
Different values, depending on the function.
}
\keyword{ts}
