\name{testRankList}
\docType{data}
\alias{testRankList}
\title{testRankList}
\usage{
data(testRankList)
}
\description{
Example data to test Apoderoides. testRankList is a list of two elements. The first element is the tip label of testTree, and the second element is corresponding family names of the tips.
}
\keyword{datasets}
